pyproject.toml
[project]
name = "banking-box"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg==0.29.0",
    "faker==30.3.0",
    "fastapi==0.115.0",
    "fastapi-cache2[redis]>=0.2.2",
    "greenlet==3.1.1",
    "httpx==0.27.2",
    "loguru>=0.7.3",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "sqlalchemy[asyncio]==2.0.35",
    "uvicorn[standard]==0.30.6",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
]

